pyproject.toml
[project]
name = "mutuacie10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "dotenv>=0.9.9",
    "google-genai>=1.66.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "qdrant-client>=1.17.0",
]
